+++ CHANGES.md
@@ -1,6 +1,13 @@
 twilio-java changelog
 =====================
 
+[2019-01-11] Version 7.34.1
+----------------------------
+**Verify**
+- Add `lookup` information in the response when creating a new verification (depends on the LookupEnabled flag being enabled at the service level)
+- Add `VerificationSid` optional parameter on Verification check.
+
+
 [2019-01-10] Version 7.34.0
 ----------------------------
 **Chat**